Clean up the validation runs from the previous day
Prior to this change, all validation runs were being stored in the db.
However validation runs aren't really needed to be kept for long as
they're just there to facilitate querying for validation.

This change deletes validation runs that are older than a day whenever a
validation run is queried for. The reason this is needed here is due to
not having any concept of a background processor. This is a poor mans
version of it.